date picker partially completed.
