+++ package.json
@@ -1,8 +1,9 @@
 {
   "name": "make-my-trip",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
+    "@datepicker-react/styled": "^2.8.4",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
