+++ src/components/SearchBox/SearchBox.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import styledComponents from 'styled-components'
+// import DatePickerWidget from '../../widgets/searchBox/DatePickerWidget'
 import DateWidget from '../../widgets/searchBox/DateWidget'
 import LocationWidget from '../../widgets/searchBox/LocationWidget'
 import TravellerWidget from '../../widgets/searchBox/TravellerWidget'
@@ -69,11 +70,14 @@ const SearchBox = () => {
 
       <DateWidget 
       label="Departure"
+      date={new Date()}
       widthValue="158" />
 
       <DateWidget 
       label="Return"
+      date={new Date()}
       widthValue="158" />
+
       <TravellerWidget />
     </SearchBoxDiv>
   )
